REV: Revert PR #958 (#1063)
* Revert "fix(parachute): gate trigger u_dot on arity instead of parameter names"

This reverts commit 6c891c97eee11435acbb0232a4a2639eb85271f1.

* Revert "MNT: refactor parachute implementation (#958)"

This reverts commit b516ace01d55821c2fe2395674a8d10052035d2e.

* REV: Fix merge errors

* MNT: remove last mentions of hemisoherical chute
