+++ docs/notebooks/monte_carlo_analysis/monte_carlo_sensitivity_simulation.py
@@ -11,14 +11,7 @@
 import datetime
 
 # %%
-from rocketpy import (
-    Environment,
-    Flight,
-    HemisphericalParachute,
-    MonteCarlo,
-    Rocket,
-    SolidMotor,
-)
+from rocketpy import Environment, Flight, MonteCarlo, Rocket, SolidMotor
 from rocketpy.stochastic import (
     StochasticEnvironment,
     StochasticFlight,
@@ -142,7 +135,7 @@ fin_set = rocket.add_trapezoidal_fins(
 tail = rocket.add_tail(
     top_radius=0.0635, bottom_radius=0.0435, length=0.060, position=-1.194656
 )
-Main = HemisphericalParachute(
+Main = rocket.add_parachute(
     "Main",
     cd_s=analysis_parameters["parachutes_main_cd_s"]["mean"],
     lag=analysis_parameters["parachutes_main_lag"]["mean"],
@@ -151,7 +144,7 @@ Main = HemisphericalParachute(
     noise=(0, 8.3, 0.5),
 )
 
-Drogue = HemisphericalParachute(
+Drogue = rocket.add_parachute(
     "Drogue",
     cd_s=analysis_parameters["parachutes_drogue_cd_s"]["mean"],
     lag=analysis_parameters["parachutes_drogue_lag"]["mean"],
@@ -159,9 +152,6 @@ Drogue = HemisphericalParachute(
     sampling_rate=105,
     noise=(0, 8.3, 0.5),
 )
-
-rocket.add_parachute(parachute=Main)
-rocket.add_parachute(parachute=Drogue)
 
 # Flight
 test_flight = Flight(
