+++ docs/reference/classes/Parachute.rst
@@ -3,9 +3,3 @@ Parachute Class
 
 .. autoclass:: rocketpy.Parachute
    :members:
-
-HemisphericalParachute Class
-----------------------------
-
-.. autoclass:: rocketpy.HemisphericalParachute
-   :members:
