+++ docs/user/rocket/rocket_usage.rst
@@ -290,25 +290,12 @@ Optionally, we can also define:
 - The parachute trigger system lag ``lag``.
 - The parachute trigger system noise ``noise``.
 
-.. note::
-
-    Since v1.13.0, :class:`~rocketpy.Parachute` is an **abstract base class**
-    and can no longer be instantiated directly. Instead, instantiate a concrete
-    parachute model such as :class:`~rocketpy.HemisphericalParachute` (used
-    below), which derives its geometry-dependent quantities (e.g. the added
-    mass during descent) from the parachute ``radius`` and ``height``. As a
-    convenience shortcut, ``Rocket.add_parachute(...)`` can still be called with
-    keyword arguments (``name``, ``cd_s``, ``trigger``, ...) and will build a
-    hemispherical parachute for you.
-
 Lets add two parachutes to the rocket, one that will be deployed at
 apogee and another that will be deployed at 800 meters above ground level:
 
 .. jupyter-execute::
 
-    from rocketpy import HemisphericalParachute
-
-    main = HemisphericalParachute(
+    main = calisto.add_parachute(
         name="Main",
         cd_s=10.0,
         trigger=800,
@@ -320,7 +307,7 @@ apogee and another that will be deployed at 800 meters above ground level:
         porosity=0.0432,
     )
 
-    drogue = HemisphericalParachute(
+    drogue = calisto.add_parachute(
         name="Drogue",
         cd_s=1.0,
         trigger="apogee",
@@ -331,8 +318,6 @@ apogee and another that will be deployed at 800 meters above ground level:
         height=1.5,
         porosity=0.0432,
     )
-    calisto.add_parachute(parachute = main)
-    calisto.add_parachute(parachute = drogue)
 
 .. seealso::
