+++ rocketpy/plots/flight_plots.py
@@ -1339,84 +1339,6 @@ class _FlightPlots:
             plt.grid(True)
             show_or_save_plot()
 
-    def parachutes_info(self, parachute_name="all"):
-        """Plots parachute dynamic information.
-        This function plots the dynamic relevant information to each parachute.
-        Different parachute models have different dynamic variables. It is
-        assumed that the 'parachutes_info' members have the dynamic variables
-        together with their respective time
-
-        Parameters
-        ----------
-        parachute_name : str | optional
-            The parachute name to display information. Default is 'all', in
-            which case information about all parachutes are plotted.
-
-        Returns
-        -------
-        None
-        """
-        # Parachute dynamic information (e.g. drag) is saved during
-        # post-processing, which is evaluated lazily. Accessing a post-processed
-        # variable forces it to run, so this plot works even when called before
-        # any other post-processed variable has been accessed.
-        _ = self.flight.ax
-
-        if not getattr(self.flight, "parachutes_info", None):
-            print("\nFlight has no parachute dynamic information available.")
-            return
-
-        if parachute_name == "all":
-            items = list(self.flight.parachutes_info.items())
-        elif parachute_name in self.flight.parachutes_info:
-            items = [(parachute_name, self.flight.parachutes_info[parachute_name])]
-        else:
-            print(
-                f"\nNo dynamic information available for parachute "
-                f"'{parachute_name}'. It may not have been deployed during the "
-                f"flight. Available parachutes: "
-                f"{list(self.flight.parachutes_info.keys())}."
-            )
-            return
-
-        # Gather every dynamic variable across the selected parachutes so each
-        # one is drawn on a single figure with all parachutes overlaid.
-        variables = []
-        for _, parachute_variables in items:
-            for variable in parachute_variables:
-                if variable != "t" and variable not in variables:
-                    variables.append(variable)
-
-        for variable in variables:
-            self.__plot_parachute_variable(variable, items)
-
-    # Axis label (and unit) for each known parachute dynamic variable.
-    __parachute_variable_labels = {"drag": ("Drag Force", "N")}
-
-    def __plot_parachute_variable(self, variable, items):
-        """Plot a single parachute dynamic variable (e.g. drag) for every
-        parachute in ``items`` overlaid on one figure, one colour each."""
-        label, unit = self.__parachute_variable_labels.get(
-            variable, (variable.replace("_", " ").title(), "")
-        )
-        ylabel = f"{label} ({unit})" if unit else label
-
-        plt.figure(figsize=(9, 4))
-        for name, parachute_variables in items:
-            if variable not in parachute_variables:
-                continue
-            plt.plot(
-                parachute_variables["t"],
-                parachute_variables[variable],
-                label=name,
-            )
-        plt.title(f"Parachute {label} vs Time")
-        plt.xlabel("Time (s)")
-        plt.ylabel(ylabel)
-        plt.legend()
-        plt.grid(True)
-        show_or_save_plot()
-
     def all(self):  # pylint: disable=too-many-statements
         """Prints out all plots available about the Flight.
 
@@ -1461,4 +1383,3 @@ class _FlightPlots:
         print("\n\nRocket and Parachute Pressure Plots\n")
         self.pressure_rocket_altitude()
         self.pressure_signals()
-        self.parachutes_info()
