+++ tests/acceptance/test_bella_lui_rocket.py
@@ -6,14 +6,7 @@ import matplotlib as mpl
 import numpy as np
 from scipy.signal import savgol_filter
 
-from rocketpy import (
-    Environment,
-    Flight,
-    Function,
-    HemisphericalParachute,
-    Rocket,
-    SolidMotor,
-)
+from rocketpy import Environment, Flight, Function, Rocket, SolidMotor
 
 mpl.rc("figure", max_open_warning=0)  # Prevent matplotlib warnings
 
@@ -139,7 +132,7 @@ def test_bella_lui_rocket_data_asserts_acceptance():
         # activate drogue when vz < 0 m/s.
         return True if y[5] < 0 else False
 
-    drogue = HemisphericalParachute(
+    BellaLui.add_parachute(
         "Drogue",
         cd_s=parameters.get("CdS_drogue")[0],
         trigger=drogue_trigger,
@@ -147,7 +140,6 @@ def test_bella_lui_rocket_data_asserts_acceptance():
         lag=parameters.get("lag_rec")[0],
         noise=(0, 8.3, 0.5),
     )
-    BellaLui.add_parachute(parachute=drogue)
 
     # Define aerodynamic drag coefficients
     power_off_drag_by_mach = Function(
